refactor: extract random policy to separate file.
